apply suggestions suggested by the amazing reviewers
