+++ test/hilbert/test_hilbert.py
@@ -16,7 +16,14 @@ import itertools
 import netket as nk
 import numpy as np
 import pytest
-from netket.hilbert import Spin, Fock, CustomHilbert, Qubit, DoubledHilbert
+from netket.hilbert import (
+    Spin,
+    Fock,
+    CustomHilbert,
+    Qubit,
+    DoubledHilbert,
+    DiscreteHilbert,
+)
 
 import jax
 import jax.numpy as jnp
@@ -104,8 +111,8 @@ hilberts["DoubledHilbert[CustomHilbert]"] = DoubledHilbert(
 )
 def test_consistent_size(hi):
     assert hi.size > 0
-    assert hi.local_size > 0
-    if hi.is_discrete:
+    if isinstance(hi, DiscreteHilbert):
+        assert hi.local_size > 0
         assert len(hi.local_states) == hi.local_size
         for state in hi.local_states:
             assert np.isfinite(state).all()
@@ -119,7 +126,7 @@ def test_random_states(hi):
     assert hi.local_size > 0
     assert len(hi.local_states) == hi.local_size
 
-    if hi.is_discrete:
+    if isinstance(hi, DiscreteHilbert):
         local_states = hi.local_states
         for i in range(100):
             rstate = hi.random_state(jax.random.PRNGKey(i * 14))
@@ -148,7 +155,7 @@ def test_flip_state(hi):
     rng = nk.jax.PRNGSeq(1)
     N_batches = 20
 
-    if hi.is_discrete:
+    if isinstance(hi, DiscreteHilbert):
         local_states = hi.local_states
         states = hi.random_state(rng.next(), N_batches)
 
@@ -182,7 +189,7 @@ def test_random_states_legacy(hi):
     assert hi.local_size > 0
     assert len(hi.local_states) == hi.local_size
 
-    if hi.is_discrete:
+    if isinstance(hi, DiscreteHilbert):
         rstate = np.zeros(hi.size)
         local_states = hi.local_states
         for i in range(100):
